pagamento/src/main/java/com/itau/pagamento/controller/ClienteController.java
package com.itau.pagamento.controller;

import com.itau.pagamento.model.Cliente;
import com.itau.pagamento.model.Cobranca;
import com.itau.pagamento.service.ClienteService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/clientes")
public class ClienteController {

    @Autowired
    private ClienteService clienteService;

    @PostMapping
    public Cliente criarCliente(@RequestBody Cliente cliente) {
        return clienteService.criarCliente(cliente);
    }
}
